Ideation for a new Rust visualisation library
Prior to this change, I was frustrated with the existing solutions for
data visualisation in multiplatform Rust programs. My experience with
the plotters crate in combination with the wgpu and winit crates was
filled with crashes from plotters backends, missing text rendering and
performance degradations.

This change introduces the starter files from previous experiments with
wgpu (flake.nix, dot files, etc.) and a comprehensive set of documents
containing the initial brainstorming work, market positioning, technical
explorations and so forth.

One notable omission of the examination of exiting libraries is the
rustplotlib crate. This was excluded as it doesn't appear to cater to
dynamic, real-time data.