Consolidate and elaborate documentation
Prior to this change, the inception phase created a lot of churn over
the documents, which led to some documents to contain out-dated and
misleading information.

This change removes many of the inception documents and consolitates
them to a few easy to follow documents. A README is introduced at the
root of the project that references these new documents.

Elaboration has begun with the introduction of the docs/planning
directory. The documents here detail each of the proposed phases from
the earlier documents, breaking them down into initiatives. The
initiatives are then broken down into stories, each with their own
detailed document under docs/planning/stories.